Refactor file naming
