+++ waitlst/src/components/LandingAnalyticsChart.vue
