--- waitlst/src/views/HomeView.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import HomeMainContent from '@/components/HomeMainContent.vue'
-import HomeNavBar from '@/components/HomeNavBar.vue'
-</script>
-
-<template>
-  <div class="grid gap-16 sm:gap-24 pb-16">
-    <HomeNavBar />
-    <HomeMainContent />
-  </div>
-</template>
